Attempt to not revert patches that have been reverted previously
This variant operates on reverts instead of cherry picks. See the
previous commit for an overview of the method and design choices.

Reverts are compared by creating a patch that is the revert of the
desired commit, of which the stable patch-id hash is used to compare
against the commit range. Notable is the mnemonic prefix (a/ and b/)
which are swapped in that revert as well, resulting in mismatched
hashes. This is addressed by showing the patch instead with
diff.noprefix=true.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>