Attempt to not pick commits that have been picked previously
repo_update is currently not reentrant: If a commit is picked once it
cannot be picked a second time as that either results in no changes or a
bunch of merge conflicts.

Generate stable hashes with patch-id that are invariant to line number
changes, white space and file ordering, and use those to compare with
existing patches on top of the selected manifest tag. Skip the pick on
matching hashes.

Note that this intentionally does not check if the patch is found
anywhere else in the tree: It might have been picked and reverted
multiple times in AOSP. Furthermore, if the patch eventually ends up on
a tag we want to be alerted too and remove it.
This obviously implies that we should not be applying and reverting our
own patches either.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>