ipacfg-mgr: Update to newer revision for matching diff context
We will be comparing picked commits/reverts and silently skip them if
already applied (only if applied on top of the selected manifest tag),
but for that both diff and context must be identical (even if using
patch-id).

The ipacfg-mgr patch has been re-pushed on top of the current tag to
have this matching context again.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>